cleaned tests for sym_wann
